Add getCharacterStream implementation
I noticed getting data out of SQLite for larger ResultSets ends up being very
slow (profiled at 7ms per row).  jstack in mixed mode indicates the call to
create a new byte buffer in the utf8BytesToJavaByteArray method is the source
of the slowness.

To get around this I've implemented getCharacterStream which will allow me
to pass a buffer into the JNI code and have SQLite copy the data into it.  The
Reader interface also allows copying only a part of the data at once.

I've added tests for this and ensured the behavior from #375 was preserved
(the Reader must be null if the column is null).