Remove copy from JNI getString
Getting data out of SQLite has involved 2 copies: 1 in JNI-land (allocating a
byte buffer to return to Java) and 1 in Java-land (creating a String instance).

One of the maintainers of the xerial/sqlite-jdbc library mentioned that we
could do away with the first copy by using a direct ByteBuffer.  This ended up
being straightforward, though I'm not sure what performance impact it will
have.  This change should be safe since as soon as the direct ByteBuffer makes
it into Java-land, it is converted into a string (and so copy #2 is performed).
So the underlying memory won't be reclaimed by SQLite while our application is
using it.

The benefits of this approach vs my other one is that we don't need to have
SQLite convert to UTF-16 under the hood so we can roughly get the same benefits
without futzing with the Java character reader.